Fix crash due to removed samples, after initial scan
When two or more samples had different name, but identical content,
only single file was stored in outputDirectory, but two references
were made in the samples list. After initial scan the file might
end up removed, but because of the second reference we would still
try to access the file.

This fix will prevent crash, but it does not fix the root cause.